refactor: split task cli command
