+++ cli/kent/main.go
@@ -123,7 +123,7 @@ func rootCommand(args []string, stdin io.Reader, stdout io.Writer, stderr io.Wri
 	if len(args) > 0 && args[0] == "workflow" {
 		return workflowSubcommand(args[1:], stdout, stderr)
 	}
-	if len(args) > 0 && args[0] == "task" {
+	if len(args) > 0 && (args[0] == "task" || args[0] == "tasks") {
 		return taskSubcommand(args[1:], stdout, stderr)
 	}
