+++ cli/kent/task_input_test.go
@@ -1,0 +1,9 @@
+package main
+
+import "testing"
+
+func TestReadTaskBodyFlagRequiresInlineOrFileBody(t *testing.T) {
+	if _, err := readTaskBodyFlag(" \t\n", ""); err == nil {
+		t.Fatal("expected missing body flags to fail")
+	}
+}
