+++ cli/kent/task_resolution.go
@@ -1,0 +1,85 @@
+package main
+
+import (
+	"context"
+	"database/sql"
+	"errors"
+	"fmt"
+	"strings"
+
+	"core/shared/config"
+	"core/shared/serverapi"
+)
+
+func workflowTaskListForProject(ctx context.Context, cfg config.App, remote workflowCommandRemote, projectRef string, req serverapi.WorkflowTaskListRequest) (serverapi.WorkflowTaskListResponse, error) {
+	projectID, err := resolveWorkflowProjectID(ctx, cfg, remote, projectRef)
+	if err != nil {
+		return serverapi.WorkflowTaskListResponse{}, err
+	}
+	req.ProjectID = projectID
+	rpcCtx, cancel := context.WithTimeout(ctx, workflowCommandTimeout)
+	defer cancel()
+	return remote.ListWorkflowTasks(rpcCtx, req)
+}
+
+func resolveWorkflowTaskID(ctx context.Context, cfg config.App, remote workflowCommandRemote, projectRef string, ref string) (string, error) {
+	trimmed := strings.TrimSpace(ref)
+	if trimmed == "" {
+		return "", errors.New("task id is required")
+	}
+	if strings.HasPrefix(trimmed, "task-") {
+		rpcCtx, cancel := context.WithTimeout(ctx, workflowCommandTimeout)
+		defer cancel()
+		resp, err := remote.GetWorkflowTask(rpcCtx, serverapi.WorkflowTaskGetRequest{TaskID: trimmed})
+		if err != nil {
+			return "", err
+		}
+		return resp.Task.Summary.ID, nil
+	}
+	projectID, err := resolveWorkflowProjectID(ctx, cfg, remote, projectRef)
+	if err != nil {
+		return "", err
+	}
+	detail, err := getWorkflowTaskByProjectShortID(ctx, remote, projectID, trimmed)
+	if err != nil {
+		if errors.Is(err, serverapi.ErrWorkflowTaskNotFound) || errors.Is(err, sql.ErrNoRows) {
+			return "", fmt.Errorf("task %q not found in project %s", trimmed, projectID)
+		}
+		return "", err
+	}
+	return detail.Summary.ID, nil
+}
+
+func isWorkflowTaskNotFound(err error) bool {
+	return errors.Is(err, sql.ErrNoRows) || errors.Is(err, serverapi.ErrWorkflowTaskNotFound)
+}
+
+func getWorkflowTaskByID(ctx context.Context, remote workflowCommandRemote, taskID string) (serverapi.WorkflowTaskDetail, error) {
+	rpcCtx, cancel := context.WithTimeout(ctx, workflowCommandTimeout)
+	defer cancel()
+	resp, err := remote.GetWorkflowTask(rpcCtx, serverapi.WorkflowTaskGetRequest{TaskID: taskID})
+	if err != nil {
+		return serverapi.WorkflowTaskDetail{}, err
+	}
+	return resp.Task, nil
+}
+
+func getWorkflowTaskByProjectShortID(ctx context.Context, remote workflowCommandRemote, projectID string, shortID string) (serverapi.WorkflowTaskDetail, error) {
+	rpcCtx, cancel := context.WithTimeout(ctx, workflowCommandTimeout)
+	defer cancel()
+	resp, err := remote.GetWorkflowTask(rpcCtx, serverapi.WorkflowTaskGetRequest{ProjectID: projectID, ShortID: shortID})
+	if err != nil {
+		return serverapi.WorkflowTaskDetail{}, err
+	}
+	return resp.Task, nil
+}
+
+func getWorkflowTaskByShortID(ctx context.Context, remote workflowCommandRemote, shortID string) (serverapi.WorkflowTaskDetail, error) {
+	rpcCtx, cancel := context.WithTimeout(ctx, workflowCommandTimeout)
+	defer cancel()
+	resp, err := remote.GetWorkflowTask(rpcCtx, serverapi.WorkflowTaskGetRequest{ShortID: shortID})
+	if err != nil {
+		return serverapi.WorkflowTaskDetail{}, err
+	}
+	return resp.Task, nil
+}
